2_IMU_improved_20250812/requirements.txt
# IMU改良モデル - 必要パッケージ
# Python 3.10+

# データ処理
numpy==1.26.4
pandas==2.2.0
polars==0.20.3
scipy==1.12.0

# 機械学習
scikit-learn==1.4.0
lightgbm==4.3.0
xgboost==2.0.3

# 設定管理
pyyaml==6.0.1

# 可視化（オプション）
matplotlib==3.8.2
seaborn==0.13.2

# ユーティリティ
joblib==1.3.2
tqdm==4.66.1

# Kaggle評価API（Kaggle環境のみ）
# kaggle-evaluation